refactor: Extract a cloud module to deal with cloud paths
